==> Tbrain_credit_card_competition/Tbrian_BryanChi_solution-master/1_Cleaning.ipynb <==
### This notebook read all query-log data
# 1. convert timestamp to tm_mon tm_day tm_hour tm_min tm_sec tm_wday(ith week)
# 2. output is march, april, may, contain each month data

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc
import time
from itertools import product

%matplotlib inline

#first let's create daily df
daily = list(np.arange(31)+1)
daily = [str(0)+str(k) if k<10 else str(k) for k in daily]
month = ['03','04','05']

March = pd.DataFrame()
month = '03'
for day in daily:
    print(' day '+str(day),end='')
    curquery = pd.read_csv('query_log/'+month+day+'.csv',header=None,dtype={0:str,1:str,2:int,3:str})
    curquery.columns = ['file_id','customer_id','timestamp','product_id']
    March = pd.concat([March,curquery],axis=0)

March.shape

March.to_hdf('march.h5','df')
#pd.read_hdf('march.h5', 'df')

April = pd.DataFrame()
month = '04'
for day in daily[:-1]:
    print(' day '+str(day),end='')
    curquery = pd.read_csv('query_log/'+month+day+'.csv',header=None,dtype={0:str,1:str,2:int,3:str})
    curquery.columns = ['file_id','customer_id','timestamp','product_id']
    April = pd.concat([April,curquery],axis=0)

April.to_hdf('april.h5','df')

April.shape

May = pd.DataFrame()
month = '05'
for day in daily:
    print(' day '+str(day),end='')
    curquery = pd.read_csv('query_log/'+month+day+'.csv',header=None,dtype={0:str,1:str,2:int,3:str})
    curquery.columns = ['file_id','customer_id','timestamp','product_id']
    May = pd.concat([May,curquery],axis=0)

May.shape

May.to_hdf('may.h5','df')

March = pd.read_hdf('March.h5', 'df')
March = March.reset_index(drop=True)

def convertime(x):
    return time.gmtime(x)
def to_tm_mon(x):
    return x.tm_mon
def to_tm_day(x):
    return x.tm_mday
def to_tm_hour(x):
    return x.tm_hour
def to_tm_min(x):
    return x.tm_min
def to_tm_sec(x):
    return x.tm_sec
def to_tm_wday(x):
    return x.tm_wday

March['strtime']=March['timestamp'].apply(convertime)
March['tm_mon']=March['strtime'].apply(to_tm_mon)
March['tm_day']=March['strtime'].apply(to_tm_day)
March['tm_hour']=March['strtime'].apply(to_tm_hour)
March['tm_min']=March['strtime'].apply(to_tm_min)
March['tm_sec']=March['strtime'].apply(to_tm_sec)
March['tm_wday']=March['strtime'].apply(to_tm_wday)

March.drop('strtime',axis=1,inplace=True)
March.head(3)

March.to_hdf('march.h5','df')

April = pd.read_hdf('april.h5', 'df')

April = April.reset_index(drop=True)
April['strtime']=April['timestamp'].apply(convertime)
April['tm_mon']=April['strtime'].apply(to_tm_mon)
April['tm_day']=April['strtime'].apply(to_tm_day)
April['tm_hour']=April['strtime'].apply(to_tm_hour)
April['tm_min']=April['strtime'].apply(to_tm_min)
April['tm_sec']=April['strtime'].apply(to_tm_sec)
April['tm_wday']=April['strtime'].apply(to_tm_wday)
April.drop('strtime',axis=1,inplace=True)
April.to_hdf('april.h5','df')

May = pd.read_hdf('may.h5', 'df')
May = May.reset_index(drop=True)
May['strtime']=May['timestamp'].apply(convertime)
May['tm_mon']=May['strtime'].apply(to_tm_mon)
May['tm_day']=May['strtime'].apply(to_tm_day)
May['tm_hour']=May['strtime'].apply(to_tm_hour)
May['tm_min']=May['strtime'].apply(to_tm_min)
May['tm_sec']=May['strtime'].apply(to_tm_sec)
May['tm_wday']=May['strtime'].apply(to_tm_wday)
May.drop('strtime',axis=1,inplace=True)
May.to_hdf('may.h5','df')
