==> Tbrain_credit_card_competition/Tbrian_BryanChi_solution-master/3_basic_feature2_Cfeat.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc

%matplotlib inline

day_cnt = pd.read_csv('./day_cnt.csv')
day_cnt.head(3)

file_df = pd.DataFrame(day_cnt.file_id.unique().tolist())
file_df.columns = ['file_id']
file_df.head(3)

np.sum(file_df.file_id == day_cnt.file_id)

file_df['first_day']=day_cnt.iloc[:,1:].apply(pd.Series.first_valid_index,axis=1).astype(int)
file_df['last_day']=day_cnt.iloc[:,1:].apply(pd.Series.last_valid_index,axis=1).astype(int)

file_df.head(3)

for i in range(92):
    i=i+1
    print(i,end=',')
    
    file_df.loc[file_df['first_day']==i,'C1']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i)]
    if i+1<93:
        file_df.loc[file_df['first_day']==i,'C2']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+1)]
    if i+2<93:
        file_df.loc[file_df['first_day']==i,'C3']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+2)]
    if i+3<93:
        file_df.loc[file_df['first_day']==i,'C4']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+3)]
    if i+4<93:
        file_df.loc[file_df['first_day']==i,'C5']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+4)]
    if i+5<93:
        file_df.loc[file_df['first_day']==i,'C6']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+5)]
file_df.head(3)

for i in range(92):
    i=i+1
    print(i,end=',')
    
    file_df.loc[file_df['last_day']==i,'C_last']=day_cnt.loc[file_df['last_day']==i,str(i)]
file_df.head(3)

file_df['C_sum'] = file_df.iloc[:,3:9].sum(axis=1)

file_df['Cper1'] = (file_df['C1']/file_df['C_sum'])*100
file_df['Cper2'] = (file_df['C2']/file_df['C_sum'])*100
file_df['Cper3'] = (file_df['C3']/file_df['C_sum'])*100
file_df['Cper4'] = (file_df['C4']/file_df['C_sum'])*100
file_df['Cper5'] = (file_df['C5']/file_df['C_sum'])*100
file_df['Cper6'] = (file_df['C6']/file_df['C_sum'])*100
file_df['Cper_last'] = (file_df['C_last']/file_df['C_sum'])*100
file_df.head(3)

for i in range(92):
    i=i+1
    day_cnt[str(i)]=day_cnt[str(i)]/day_cnt[str(i)].median()
day_cnt.head(3)

for i in range(92):
    i=i+1
    print(i,end=',')
    
    file_df.loc[file_df['first_day']==i,'Cm1']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i)]
    if i+1<93:
        file_df.loc[file_df['first_day']==i,'Cm2']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+1)]
    if i+2<93:
        file_df.loc[file_df['first_day']==i,'Cm3']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+2)]
    if i+3<93:
        file_df.loc[file_df['first_day']==i,'Cm4']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+3)]
    if i+4<93:
        file_df.loc[file_df['first_day']==i,'Cm5']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+4)]
    if i+5<93:
        file_df.loc[file_df['first_day']==i,'Cm6']=day_cnt.loc[pd.notnull(day_cnt)[str(i)],str(i+5)]
file_df.head(3)

for i in range(92):
    i=i+1
    print(i,end=',')
    
    file_df.loc[file_df['last_day']==i,'Cm_last']=day_cnt.loc[file_df['last_day']==i,str(i)]
file_df.head(3)

file_df['Cm_sum'] = file_df.iloc[:,18:24].sum(axis=1)

file_df['Cmper1'] = (file_df['Cm1']/file_df['Cm_sum'])*100
file_df['Cmper2'] = (file_df['Cm2']/file_df['Cm_sum'])*100
file_df['Cmper3'] = (file_df['Cm3']/file_df['Cm_sum'])*100
file_df['Cmper4'] = (file_df['Cm4']/file_df['Cm_sum'])*100
file_df['Cmper5'] = (file_df['Cm5']/file_df['Cm_sum'])*100
file_df['Cmper6'] = (file_df['Cm6']/file_df['Cm_sum'])*100
file_df['Cmper_last'] = (file_df['Cm_last']/file_df['Cm_sum'])*100
file_df.head(3)

# now we have c1~c6, cper1~cper6, cm1~cm6, cmper1~cmper6
# take min max mean median std 
file_df['C_mean'] = np.nanmean(file_df[['C1','C2','C3','C4','C5','C6']].values,axis=1)
file_df['C_median'] = np.nanmedian(file_df[['C1','C2','C3','C4','C5','C6']].values,axis=1)
file_df['C_std']= np.nanstd(file_df[['C1','C2','C3','C4','C5','C6']].values,axis=1)
file_df['C_min']= np.nanmin(file_df[['C1','C2','C3','C4','C5','C6']].values,axis=1)
file_df['C_max']= np.nanmax(file_df[['C1','C2','C3','C4','C5','C6']].values,axis=1)

file_df.head(3)

file_df['Cm_mean'] = np.nanmean(file_df[['Cm1','Cm2','Cm3','Cm4','Cm5','Cm6']].values,axis=1)
file_df['Cm_median'] = np.nanmedian(file_df[['Cm1','Cm2','Cm3','Cm4','Cm5','Cm6']].values,axis=1)
file_df['Cm_std']= np.nanstd(file_df[['Cm1','Cm2','Cm3','Cm4','Cm5','Cm6']].values,axis=1)
file_df['Cm_min']= np.nanmin(file_df[['Cm1','Cm2','Cm3','Cm4','Cm5','Cm6']].values,axis=1)
file_df['Cm_max']= np.nanmax(file_df[['Cm1','Cm2','Cm3','Cm4','Cm5','Cm6']].values,axis=1)

file_df.head(3)

file_df['Cmper_mean'] = np.nanmean(file_df[['Cmper1','Cmper2','Cmper3','Cmper4','Cmper5','Cmper6']].values,axis=1)
file_df['Cmper_median'] = np.nanmedian(file_df[['Cmper1','Cmper2','Cmper3','Cmper4','Cmper5','Cmper6']].values,axis=1)
file_df['Cmper_std']= np.nanstd(file_df[['Cmper1','Cmper2','Cmper3','Cmper4','Cmper5','Cmper6']].values,axis=1)
file_df['Cmper_min']= np.nanmin(file_df[['Cmper1','Cmper2','Cmper3','Cmper4','Cmper5','Cmper6']].values,axis=1)
file_df['Cmper_max']= np.nanmax(file_df[['Cmper1','Cmper2','Cmper3','Cmper4','Cmper5','Cmper6']].values,axis=1)

file_df.head(3)

file_df['Cper_mean'] = np.nanmean(file_df[['Cper1','Cper2','Cper3','Cper4','Cper5','Cper6']].values,axis=1)
file_df['Cper_median'] = np.nanmedian(file_df[['Cmper1','Cper2','Cper3','Cper4','Cper5','Cper6']].values,axis=1)
file_df['Cper_std']= np.nanstd(file_df[['Cper1','Cper2','Cper3','Cper4','Cper5','Cper6']].values,axis=1)
file_df['Cper_min']= np.nanmin(file_df[['Cper1','Cper2','Cper3','Cper4','Cper5','Cper6']].values,axis=1)
file_df['Cper_max']= np.nanmax(file_df[['Cper1','Cper2','Cper3','Cper4','Cper5','Cper6']].values,axis=1)

file_df.head(3)

file_df.fillna(0,inplace=True)

file_df['C1_C2'] = file_df['C1'] - file_df['C2']
file_df['Cfirst_Clast'] = file_df['C1'] - file_df['C_last']
file_df['Cmax_Cmin'] = file_df['C_max'] - file_df['C_min']

file_df['Cm1_Cm2'] = file_df['Cm1'] - file_df['Cm2']
file_df['Cmfirst_Cmlast']= file_df['Cm1'] - file_df['Cm_last']
file_df['Cmmax_Cmmin'] = file_df['Cm_max'] - file_df['Cm_min']

file_df['Cper1_Cper2'] = file_df['Cper1'] - file_df['Cper2']
file_df['Cperfirst_Cperlast'] = file_df['Cper1'] - file_df['Cper_last']
file_df['Cpermax_Cpermin'] = file_df['Cper_max'] - file_df['Cper_min']

file_df['Cmper1_Cmper2'] = file_df['Cmper1'] - file_df['Cmper2']
file_df['Cmperfirst_Cmperlast'] = file_df['Cmper1'] - file_df['Cmper_last']
file_df['Cmpermax_Cmpermin'] = file_df['Cmper_max'] - file_df['Cmper_min']

file_df.drop('target',axis=1,inplace=True)

file_df.to_csv('file_df.csv',index=False)
