==> Tbrain_credit_card_competition/Tbrian_BryanChi_solution-master/4_prod_cust_cnt.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc

%matplotlib inline

product_cnt = pd.read_csv('product_cnt.csv')
cust_prod = pd.read_csv('cust_prod.csv')
train = pd.read_csv('training-set.csv',header=None)
test = pd.read_csv('testing-set.csv',header=None)
train.columns = ['file_id','target']
test.columns = ['file_id','target']

product_cnt.head(3)

cust_prod.head(3)

# cust_freq, prod_freq

file_df = pd.read_csv('file_df.csv')

file_df.head(3)

file_df.columns

file_df = file_df[['file_id','C_sum']]
cust_prod = pd.merge(cust_prod,file_df,how='left',on='file_id')

cust_prod.head(3)

cust_prod['cust_freq']=cust_prod['cust_cnt']/cust_prod['C_sum']
cust_prod['prod_freq']=cust_prod['prod_cnt']/cust_prod['C_sum']

cust_prod.to_csv('cust_prod.csv',index=False)

# custxprod_cnt, custxprod_freq

file_df = pd.read_csv('file_df.csv')

file_df = file_df[['file_id','first_day','last_day']]

cust_prod = pd.read_csv('cust_prod.csv')
cust_prod.head(3)

march = pd.read_hdf('march.h5', 'df')
march['custxprod'] = march['customer_id']+march['product_id']
march = march[['file_id','custxprod','timestamp']]
march = march.groupby(['file_id','custxprod']).count().reset_index()

march.head(3)

file_list = file_df['file_id'].unique().tolist()
marfile = file_df[file_df['last_day']<32].file_id.tolist()
aprfile = file_df[(file_df['first_day']>31)&(file_df['last_day']<62)].file_id.tolist()
mayfile = file_df[(file_df['first_day']>61)&(file_df['last_day']<93)].file_id.tolist()
otherfile = list(set(file_list)-set(marfile)-set(aprfile)-set(mayfile))

len(marfile), len(aprfile), len(mayfile), len(otherfile)

march_file_df = march[march.isin(marfile).file_id]
march_other_df = march[march.isin(otherfile).file_id]
del march
gc.collect();

march_file_df = march_file_df.groupby('file_id').count().reset_index()

march_file_df=march_file_df[['file_id','custxprod']]

march_file_df.shape

april = pd.read_hdf('april.h5', 'df')
april['custxprod'] = april['customer_id']+april['product_id']
april = april[['file_id','custxprod','timestamp']]
april = april.groupby(['file_id','custxprod']).count().reset_index()

april_file_df = april[april.isin(aprfile).file_id]
april_other_df = april[april.isin(otherfile).file_id]

del april
gc.collect();

april_file_df = april_file_df.groupby('file_id').count().reset_index()
april_file_df=april_file_df[['file_id','custxprod']]

may = pd.read_hdf('may.h5', 'df')
may['custxprod'] = may['customer_id']+may['product_id']
may = may[['file_id','custxprod','timestamp']]
may = may.groupby(['file_id','custxprod']).count().reset_index()

may_file_df = may[may.isin(mayfile).file_id]
may_other_df = may[may.isin(otherfile).file_id]

del may
gc.collect();

may_file_df = may_file_df.groupby('file_id').count().reset_index()
may_file_df=may_file_df[['file_id','custxprod']]


#deal with other file
other_file = pd.concat([march_other_df,april_other_df,may_other_df])
del march_other_df, april_other_df, may_other_df
gc.collect();

other_file = other_file.groupby(['file_id','custxprod']).count().reset_index()
other_file = other_file.groupby('file_id').count().reset_index()
other_file.head(3)

other_file=other_file[['file_id','custxprod']]
data = pd.concat([march_file_df,april_file_df,may_file_df,other_file],axis=0)

data.columns = ['file_id','custxprod_cnt']
cust_prod.head(3)

cust_prod = pd.merge(cust_prod,data,how='left',on='file_id')

cust_prod.head(3)

cust_prod['custxprod_freq']=cust_prod['custxprod_cnt']/cust_prod['C_sum']

cust_prod.drop('C_sum',axis=1,inplace=True)

cust_prod.to_csv('cust_prod.csv',index=False)
