==> Tbrain_credit_card_competition/Tbrian_BryanChi_solution-master/6_more_interval_feature.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc

%matplotlib inline

# hrinte_cust

# 把小時切成 0~6, 6~12, 12~18, 18~24
#  21~ 23 + 0~3 , 3~9, 9~15, 15~21
time_interval_dict = dict()
for i in range(24):
    j = i//6 
    #j=(i-3)//6
    if j<0:
        j=3
    time_interval_dict[i]=j

march = pd.read_hdf('march.h5', 'df')

march['hour_cat'] = march['tm_hour'].map(time_interval_dict)

march = march.groupby(['file_id','customer_id','hour_cat']).count().reset_index()[['file_id','customer_id','hour_cat','timestamp']]
march['hour_cat'] = march['hour_cat']+1
march.head(3)

april = pd.read_hdf('april.h5', 'df')

april['hour_cat'] = april['tm_hour'].map(time_interval_dict)

april = april.groupby(['file_id','customer_id','hour_cat']).count().reset_index()[['file_id','customer_id','hour_cat','timestamp']]
april['hour_cat'] = april['hour_cat']+1
april.head(3)

may = pd.read_hdf('may.h5', 'df')

may['hour_cat'] = may['tm_hour'].map(time_interval_dict)

may = may.groupby(['file_id','customer_id','hour_cat']).count().reset_index()[['file_id','customer_id','hour_cat','timestamp']]
may['hour_cat'] = may['hour_cat']+1
may.head(3)

data = pd.concat([march,april,may],axis=0)
data.head(3)

data = data.groupby(['file_id','customer_id','hour_cat']).sum().reset_index()
data.head(3)

data = data.groupby(['file_id','hour_cat']).count().reset_index()[['file_id','hour_cat','timestamp']]
data.head(3)

data = data.pivot_table(values='timestamp',index=['file_id'],columns='hour_cat',aggfunc='sum').reset_index()
data = data.fillna(0)
data.columns = ['file_id', 'cchrinte1', 'cchrinte2', 'cchrinte3', 'cchrinte4']
data.head(3)

data.to_csv('hour_interval_cust_cnt.csv',index=False)

# mininte_cust

# 把分鐘切成 0~15, 15~30, 30~45, 45~60
time_interval_dict = dict()
for i in range(60):
    #j = i//6 
    j=(i)//15
    time_interval_dict[i]=j

march = pd.read_hdf('march.h5', 'df')

march['min_cat'] = march['tm_min'].map(time_interval_dict)

march = march.groupby(['file_id','customer_id','min_cat']).count().reset_index()[['file_id','customer_id','min_cat','timestamp']]
march['min_cat'] = march['min_cat']+1
march.head(3)


april = pd.read_hdf('april.h5', 'df')

april['min_cat'] = april['tm_min'].map(time_interval_dict)

april = april.groupby(['file_id','customer_id','min_cat']).count().reset_index()[['file_id','customer_id','min_cat','timestamp']]
april['min_cat'] = april['min_cat']+1
april.head(3)

may = pd.read_hdf('may.h5', 'df')

may['min_cat'] = may['tm_min'].map(time_interval_dict)

may = may.groupby(['file_id','customer_id','min_cat']).count().reset_index()[['file_id','customer_id','min_cat','timestamp']]
may['min_cat'] = may['min_cat']+1
may.head(3)

data = pd.concat([march,april,may],axis=0)
data.head(3)

data = data.groupby(['file_id','customer_id','min_cat']).sum().reset_index()
data.head(3)

data = data.groupby(['file_id','min_cat']).count().reset_index()[['file_id','min_cat','timestamp']]
data.head(3)

data = data.pivot_table(values='timestamp',index=['file_id'],columns='min_cat',aggfunc='sum').reset_index()
data = data.fillna(0)
data.columns = ['file_id', 'ccmininte1', 'ccmininte2', 'ccmininte3', 'ccmininte4']
data.head(3)

data.to_csv('min_interval_cust_cnt.csv',index=False)

# sec_cust

# 把秒切成 0~15, 15~30, 30~45, 45~60
time_interval_dict = dict()
for i in range(60):
    #j = i//6 
    j=(i)//15
    time_interval_dict[i]=j

march = pd.read_hdf('march.h5', 'df')

march['sec_cat'] = march['tm_sec'].map(time_interval_dict)

march = march.groupby(['file_id','customer_id','sec_cat']).count().reset_index()[['file_id','customer_id','sec_cat','timestamp']]
march['sec_cat'] = march['sec_cat']+1
march.head(3)

april = pd.read_hdf('april.h5', 'df')

april['sec_cat'] = april['tm_sec'].map(time_interval_dict)

april = april.groupby(['file_id','customer_id','sec_cat']).count().reset_index()[['file_id','customer_id','sec_cat','timestamp']]
april['sec_cat'] = april['sec_cat']+1
april.head(3)

may = pd.read_hdf('may.h5', 'df')

may['sec_cat'] = may['tm_sec'].map(time_interval_dict)

may = may.groupby(['file_id','customer_id','sec_cat']).count().reset_index()[['file_id','customer_id','sec_cat','timestamp']]
may['sec_cat'] = may['sec_cat']+1
may.head(3)

data = pd.concat([march,april,may],axis=0)
data.head(3)

data = data.groupby(['file_id','customer_id','sec_cat']).sum().reset_index()
data.head(3)

data = data.groupby(['file_id','sec_cat']).count().reset_index()[['file_id','sec_cat','timestamp']]
data.head(3)

data = data.pivot_table(values='timestamp',index=['file_id'],columns='sec_cat',aggfunc='sum').reset_index()
data = data.fillna(0)
data.columns = ['file_id', 'ccsecinte1', 'ccsecinte2', 'ccsecinte3', 'ccsecinte4']
data.head(3)

data.to_csv('sec_interval_cust_cnt.csv',index=False)
