==> Tbrain_credit_card_competition/Tbrian_BryanChi_solution-master/8_exp_mean_encoding_smooth.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc

%matplotlib inline

# create long representation for file and cust
#march = pd.read_hdf('./march.h5', 'df')
#march = march.groupby(['file_id','customer_id']).count().reset_index()[['file_id','customer_id','timestamp']]

#april = pd.read_hdf('./april.h5', 'df')
#april = april.groupby(['file_id','customer_id']).count().reset_index()[['file_id','customer_id','timestamp']]

#may = pd.read_hdf('./may.h5', 'df')
#may = may.groupby(['file_id','customer_id']).count().reset_index()[['file_id','customer_id','timestamp']]

#data = pd.concat([march,april,may],axis=0)

#data = data.groupby(['file_id','customer_id']).count().reset_index()[['file_id','customer_id']]
#data.head(3)

data.to_hdf('file_cust_longrep.h5','df')
data = pd.read_hdf('file_cust_longrep.h5', 'df')
data.head(3)

file_df = pd.read_csv('file_df.csv')
file_df = file_df[['file_id','first_day']]

data = pd.merge(data,file_df,on='file_id',how='left')

train = pd.read_csv('training-set.csv',header=None)
test = pd.read_csv('testing-set.csv',header=None)
train.columns = ['file_id','target']
test.columns = ['file_id','target']

### full mean code
#fullmeancode = pd.read_csv('final/forfullmeancode.csv')
#train = pd.concat([train,fullmeancode],axis=0)
### full mean code

data = pd.merge(data,train,how='left',on='file_id')
data.head(3)

data = data.sort_values(by=['first_day','file_id']).reset_index(drop=True)
data.head(3)

tr_data = data[data['first_day']<50] # 62 for predict, 50 for val
tr_data = tr_data[['file_id','customer_id','target']]
tr_data.head(3)

cumsum = tr_data.groupby('customer_id').cumsum().target - tr_data['target']
cumcnt = tr_data.groupby('customer_id').cumcount()
tr_data['mean_score'] = cumsum / cumcnt

tr_data['mean_score'].mean()

alpha = 20
gbmean = 0.006764493003441334

tr_data['mean_score'] = ((tr_data['mean_score']*cumcnt)+(gbmean*alpha))/(cumcnt+alpha)

customer_score = tr_data.groupby('customer_id').last().reset_index()[['customer_id','mean_score']]
customer_score.head(3)

customer_df = pd.read_csv('customer_df.csv')
customer_df = pd.merge(customer_df,customer_score,how='left',on='customer_id')
customer_df = customer_df[['customer_id','mean_score']]
customer_df.head(3)

#customer_df.loc[pd.isnull(customer_df).mean_score,'mean_score']=0.012656403967587571
customer_df = customer_df[customer_df['mean_score']>-1]
customer_df.shape

te_data = data[data['first_day']>49]
te_data.head(3)

te_data = te_data.sort_values(by=['first_day','file_id']).reset_index(drop=True)
te_data = te_data[['file_id','customer_id']]
te_data = pd.merge(te_data,customer_df,how='left',on='customer_id')
te_data.head(3)

data = pd.concat([tr_data,te_data],ignore_index=True)
data.head(3)

#data.loc[pd.isnull(data).mean_score,'mean_score']=0.012656403967587571
data = data[data['mean_score']>-1]
data.head(3)

data = data[['file_id','mean_score']]
exp_menc_cust = data.groupby('file_id').mean().reset_index()
exp_menc_cust.head(3)

train = pd.read_csv('training-set.csv',header=None)
test = pd.read_csv('testing-set.csv',header=None)
train.columns = ['file_id','target']
test.columns = ['file_id','target']
trte = pd.concat([train,test],axis=0)
trte = pd.merge(trte,exp_menc_cust,how='left',on='file_id')
trte = trte[['file_id','mean_score']]
trte.head(3)

trte.columns = ['file_id','cust_mean']

trte['cust_mean'].mean()

trte.loc[pd.isnull(trte)['cust_mean'],'cust_mean']=0.03535477004433718

trte.to_csv('cust_expsmo_meancode_val.csv',index=False)
trte.shape
